+++ test/list.js
@@ -18,6 +18,7 @@ var sum = hs.list.sum
 var append = hs.list.append
 var range = hs.list.range
 var filter = hs.list.filter
+var concat = hs.list.concat
 
 describe('list module', function() {
   describe('length([1,3])', function() {
@@ -122,6 +123,15 @@ describe('list module', function() {
       var appendToMyList = append([1,2])
       expect(appendToMyList).to.be.a('function')
       expect(appendToMyList([3,4])).to.be.deep.equal([1,2,3,4])
+    })
+  })
+  describe('concat([[1,2],[3,4])', function() {
+    it('returns [1,2,3,4]', function() {
+      var xs = [1,2]
+      var ys = [3,4]
+      expect(concat([xs,ys])).to.be.deep.equal([1,2,3,4])
+      expect(xs).to.be.deep.equal([1,2])
+      expect(ys).to.be.deep.equal([3,4])
     })
   })
   describe('range(1,5)', function() {