+++ hs.js
@@ -1,8 +1,12 @@
 var core = require('./lib/core')
 var base = require('./lib/base')
 var list = require('./lib/list')
+var sort = require('./lib/sort')
+
+list.n = require('./lib/list.n')
 
 exports.core = core
 exports.base = base
 exports.list = list
+exports.sort = sort