+++ test/base.js
@@ -9,6 +9,8 @@ var max = hs.base.max
 var div = hs.base.div
 var add = hs.base.add
 var subtract = hs.base.subtract
+var lessEqual = hs.base.lessEqual
+var greaterThan = hs.base.greaterThan
 
 describe('base', function() {
   describe('succ(8)', function() {
@@ -56,4 +58,35 @@ describe('base', function() {
       expect(subtractTwo(3)).to.be.equal(1)
     })
   })
+  describe('lessEqual(4, 10)', function() {
+    it('returns false', function() {
+      expect(lessEqual(4,10)).to.be.false
+    })
+  })
+  describe('lessEqual(2, 2)', function() {
+    it('returns true', function() {
+      expect(lessEqual(2,2)).to.be.true
+    })
+  })
+  describe('lessEqual(3, 1)', function() {
+    it('returns true', function() {
+      expect(lessEqual(3,1)).to.be.true
+    })
+  })
+  describe('greaterThan(4, 10)', function() {
+    it('returns true', function() {
+      expect(greaterThan(4,10)).to.be.true
+    })
+  })
+  describe('greaterThan(2, 2)', function() {
+    it('returns false', function() {
+      expect(greaterThan(2,2)).to.be.false
+    })
+  })
+  describe('greaterThan(3, 1)', function() {
+    it('returns false', function() {
+      expect(greaterThan(3,1)).to.be.false
+    })
+  })
+
 })