+++ test/list.js
@@ -4,6 +4,7 @@ var expect = require('chai').expect
 // import functions
 var hs = require('../hs')
 var add = hs.base.add
+var greaterThan = hs.base.greaterThan
 var length = hs.list.length
 var empty = hs.list.empty
 var head = hs.list.head
@@ -16,6 +17,8 @@ var reduce = hs.list.reduce
 var sum = hs.list.sum
 var append = hs.list.append
 var range = hs.list.range
+var filter = hs.list.filter
+var concat = hs.list.concat
 
 describe('list module', function() {
   describe('length([1,3])', function() {
@@ -122,6 +125,15 @@ describe('list module', function() {
       expect(appendToMyList([3,4])).to.be.deep.equal([1,2,3,4])
     })
   })
+  describe('concat([[1,2],[3,4])', function() {
+    it('returns [1,2,3,4]', function() {
+      var xs = [1,2]
+      var ys = [3,4]
+      expect(concat([xs,ys])).to.be.deep.equal([1,2,3,4])
+      expect(xs).to.be.deep.equal([1,2])
+      expect(ys).to.be.deep.equal([3,4])
+    })
+  })
   describe('range(1,5)', function() {
     it('returns [1,2,3,4,5]', function() {
       expect(range(1,5)).to.be.deep.equal([1,2,3,4,5])
@@ -140,6 +152,13 @@ describe('list module', function() {
   describe('range(1,2.5)', function() {
     it('returns [1,2]', function() {
       expect(range(1,2)).to.be.deep.equal([1,2])
+    })
+  })
+  describe('filter(greaterThan(2), [1,2,3,4])', function() {
+    it('returns [3,4]', function() {
+      var xs = [1,2,3,4]
+      expect(filter(greaterThan(2), xs)).to.be.deep.equal([3,4])
+      expect(xs).to.be.deep.equal([1,2,3,4])
     })
   })
 })